Complete Checkout Module

+++ Models/Checkouts/Checkout.cs
@@ -7,7 +7,7 @@
         public string LastName { get; set; } = string.Empty;
         public string Email { get; set; } = string.Empty;
         public string Address { get; set; } = string.Empty;
-        public float Payment { get; set; }
+        public double Payment { get; set; }
 
         // RELATION WITH TABLE CUSTOMER
         public Customer? Customers { get; set; }
